feat: create private

+++ gym-rat/src/app/api/auth/user/route.ts
@@ -1,9 +1,19 @@
 import connectMongoDB from "@/lib/mongodb";
 import UserModel, { iUser } from "@/models/userModel";
+import { getServerSession } from "next-auth";
 import { NextResponse } from "next/server";
 
 export async function POST(req: any) {
   try {
+    const session = await getServerSession(req);
+    // Check if the user is authenticated
+    console.log("session", session);
+    if (!session) {
+      return NextResponse.json(
+        { error: "User is not authenticated" },
+        { status: 401 }
+      );
+    }
     const params: iUser = await req.json(); // get params from request
     await connectMongoDB(); // connect to BD
     const oldUser = await UserModel.findOne({ email: params.email }); // check that user was created
@@ -24,12 +34,34 @@ export async function POST(req: any) {
   }
 }
 
-export async function GET() {
-  // await connectMongoDB();
+export async function GET(req: any) {
+  const session = await getServerSession(req);
+  // Check if the user is authenticated
+  console.log("session", session);
+  if (!session) {
+    return NextResponse.json(
+      { error: "User is not authenticated" },
+      { status: 401 }
+    );
+  }
+  //
+  await connectMongoDB();
+  //
   const users = await UserModel.find();
   return NextResponse.json({ users }, { status: 200 });
 }
+
 export async function DELETE(req: any) {
+  const session = await getServerSession(req);
+  // Check if the user is authenticated
+  console.log("session", session);
+  if (!session) {
+    return NextResponse.json(
+      { error: "User is not authenticated" },
+      { status: 401 }
+    );
+  }
+  //
   const id = req.nextUrl.searchParams.get("id");
   await connectMongoDB();
   await UserModel.findByIdAndDelete(id);
